+++ ReleaseNotes.md
@@ -1,3 +1,4 @@
 # Bluedot React Native Plugin release notes
 
-- Integrated with Bluedot Android SDK v17.3.0 and iOS SDK 17.1.0
+- Push Notifications feature
+- Integrated with Bluedot Android SDK v18.0.0 and iOS SDK 18.0.0